+++ src/com/brandon/snake/render/renderer/snake/HeadBlinkAnimation.java
@@ -33,7 +33,7 @@ public class HeadBlinkAnimation extends Animation {
 	}
 
 	@Override
-	protected void onUpdate(float time, float deltaTime) {
+	protected void onUpdate(long time, long deltaTime) {
 		current = 1 - current; //Toggles between 0 and 1
 		models.removeFirst();
 		models.addFirst(headModels[current]);
